integ-tests: test MPI_Send/Recv in test_mpi
