+++ tests/integration-tests/tests/common/data/mpi/mpi_submit_no_module_available.sh
@@ -5,4 +5,4 @@ rm -f /shared/mpi.out
 # -mca btl ^openib is needed to avoid to have the following warning in the stdout:
 # A high-performance Open MPI point-to-point messaging module was unable to find any relevant network interfaces:
 # Module: OpenFabrics (openib). Another transport will be used instead, although this may result in lower performance.
-mpirun --map-by ppr:1:node -mca btl ^openib "mpi_hello_world" >> /shared/mpi.out
+mpirun --map-by ppr:1:node -mca btl ^openib --timeout 10 "ring" >> /shared/mpi.out
