+++ tests/integration-tests/tests/common/data/mpi/mpi_submit_openmpi.sh
@@ -3,4 +3,4 @@ set -e
 
 rm -f /shared/mpi.out
 module load openmpi
-mpirun --map-by ppr:1:node "mpi_hello_world" >> /shared/mpi.out
+mpirun --map-by ppr:1:node --timeout 10 "ring" >> /shared/mpi.out
